+++ examples/bun/src/rpc.ts
@@ -1,8 +1,8 @@
 import { CompatibilityLevel, createClient } from "polkadot-api"
-import { WebSocketProvider } from "polkadot-api/ws-provider/node"
+import { getWsProvider } from "polkadot-api/ws-provider/node"
 import { wnd } from "@polkadot-api/descriptors"
 
-const client = createClient(WebSocketProvider("wss://westend-rpc.polkadot.io"))
+const client = createClient(getWsProvider("wss://westend-rpc.polkadot.io"))
 const testApi = client.getTypedApi(wnd)
 
 const isCompatible = (level: CompatibilityLevel) =>