+++ packages/json-rpc/ws-provider/src/web.ts
@@ -1,4 +1,4 @@
-import { getWebSocketProvider } from "./ws-provider"
+import { getInternalWsProvider } from "./ws-provider"
 
 export type { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
-export const WebSocketProvider = getWebSocketProvider(WebSocket)
+export const getWsProvider = getInternalWsProvider(WebSocket)